requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
Faker==30.8.2
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
